+++ Apps/ValidationTests/Win32/App.cpp
@@ -87,10 +87,7 @@ namespace
         Babylon::InitializeNativeEngine(*runtime, hWnd, width, height);
 
         // Initialize XMLHttpRequest plugin.
-        runtime->Dispatch([&runtime = *runtime](Napi::Env env)
-        {
-            Babylon::XMLHttpRequest::Initialize(env, runtime.RootUrl.data());
-        });
+        Babylon::InitializeXMLHttpRequest(*runtime, runtime->RootUrl.data());
 
         runtime->Dispatch([hWnd](Napi::Env env)
         {
