+++ Plugins/XMLHttpRequest/CMakeLists.txt
@@ -1,16 +1,16 @@
 set(SOURCES 
     "Include/Babylon/XMLHttpRequest.h"
     "Source/XMLHttpRequest.cpp"
+    "Source/XMLHttpRequest.h"
     "Source/XMLHttpRequest${BABYLON_NATIVE_PLATFORM}.cpp")
 
 add_library(XMLHttpRequest ${SOURCES})
 
-target_include_directories(XMLHttpRequest PRIVATE "Include/Babylon")
-target_include_directories(XMLHttpRequest INTERFACE "Include")
+target_include_directories(XMLHttpRequest PUBLIC "Include")
 
 target_link_libraries(XMLHttpRequest 
     PUBLIC JsRuntime
-    PUBLIC arcana
+    PRIVATE arcana
     PRIVATE BabylonNativeUtils)
 
 if(WINDOWS_STORE)
